Rename localstorage cache module file
