+++ src/cache/index.ts
@@ -1,4 +1,4 @@
-export * from './cache-local-storage';
+export * from './cache-localstorage';
 export * from './cache-memory';
 export * from './cache-manager';
 export { ICache, CacheKey } from './shared';
